reorganize item creation
